refactor(api): Move validator utils to validatorFactor.go
