+++ src/api/validatorFactory.go
@@ -1,23 +1,21 @@
-package conf
+package api
 
 import(
     "reflect"
-    "api"
-//    "fmt"
     )
 
 var typeMap = map[string]reflect.Type{
-    "httpCode": reflect.TypeOf(api.HttpCodeChecker{}),
+    "httpCode": reflect.TypeOf(HttpCodeChecker{}),
 }
 
-func GetValidator(typeStr string) api.ApiValidator{
+func GetValidator(typeStr string) ApiValidator{
     t:= typeMap[typeStr]
     v:= reflect.New(t).Elem()
     i:= v.Interface()
-    validator:= i.(api.ApiValidator)
+    validator:= i.(ApiValidator)
     return validator
 }
-func GetValidatorType(v api.ApiValidator) string{
+func GetValidatorType(v ApiValidator) string{
     t := reflect.TypeOf(v)
     for k,v:= range typeMap{
         if v == t{
