+++ src/conf/util.go
@@ -1,13 +1,14 @@
 package conf
 
 import(
+    "api"
     )
 
 func fromBasic(src basicConf)ApiConf{
-   valid:= GetValidator(src.ValidatorType)
+   valid:= api.GetValidator(src.ValidatorType)
    return ApiConf{src.Api, valid, src.Tags}    
 }
 func toBasic(src ApiConf)basicConf{
-    validType:= GetValidatorType(src.Validator)
+    validType:= api.GetValidatorType(src.Validator)
     return basicConf{src.Api, validType, src.Tags}
 }
